Backend/requirements.txt
absl-py==2.4.0
altair==6.0.0
asgiref==3.11.0
astunparse==1.6.3
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dash==3.3.0
dash_ag_grid==32.3.2
distlib==0.4.0
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
filelock==3.20.0
Flask==3.1.2
flatbuffers==25.12.19
fonttools==4.61.1
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
grpcio==1.78.1
h5py==3.15.1
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.13.2
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
narwhals==2.12.0
nest-asyncio==1.6.0
numpy==2.3.5
opencv-python==4.13.0.92
opt_einsum==3.4.0
optree==0.19.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
platformdirs==4.5.0
plotly==6.5.0
protobuf==6.33.2
pyarrow==22.0.0
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.37.0
requests==2.32.5
retrying==1.4.2
rich==14.3.3
rpds-py==0.30.0
scipy==1.16.3
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
sqlparse==0.5.3
streamlit==1.52.2
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
toml==0.10.2
tornado==6.5.4
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.35.4
watchdog==6.0.0
Werkzeug==3.1.3
wrapt==2.1.1
zipp==3.23.0
